https://www.projetoacbr.com.br/forum/topic/93564-acbr-9649-pix-autom%C3%A1tico-campo-loc-da-recorr%C3%AAncia-trunca-id-de-location-grande-integer-%E2%86%92-deveria-ser-int64/
git-svn-id: https://svn.code.sf.net/p/acbr/code/trunk2@47230 6e92efe7-b92a-0410-a9ec-f9e4e41bb3a6

+++ Fontes/ACBrPIXCD/ACBrPIXSchemasRec.pas
@@ -124,12 +124,12 @@ type
 
   TACBrPIXRecSolicitadaBase = class(TACBrPIXRecBase)
   private
-    fLoc: Integer;
+    fLoc: Int64;
   protected
     procedure DoWriteToJSon(AJSon: TACBrJSONObject); override;
     procedure DoReadFromJSon(AJSon: TACBrJSONObject); override;
 
-    property loc: Integer read fLoc write fLoc;
+    property loc: Int64 read fLoc write fLoc;
   public
     procedure Clear; override;
     function IsEmpty: Boolean; override;
